% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_rbmi.R
\name{a_rbmi_lsmeans}
\alias{a_rbmi_lsmeans}
\title{Formatted Analysis function which can be further customized by calling
\code{\link[rtables:make_afun]{rtables::make_afun()}} on it. It is used as \code{afun} in \code{\link[rtables:analyze]{rtables::analyze()}}.}
\usage{
a_rbmi_lsmeans(df, .in_ref_col, show_relative = c("reduction", "increase"))
}
\arguments{
\item{df}{input dataframe}

\item{.in_ref_col}{boolean variable, if reference column is specified}

\item{show_relative}{"reduction" if (\code{control - treatment}, default) or "increase"
(\code{treatment - control}) of relative change from baseline?}
}
\value{
Formatted Analysis function
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
