% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tern.rbmi-package.R
\docType{data}
\name{rbmi_test_data}
\alias{rbmi_test_data}
\title{Example dataset for \code{tern.rbmi} package. This is an pool object from the rbmi analysis, see
\code{browseVignettes(package = "tern.rbmi")}}
\format{
An object of class \code{pool} of length 5.
}
\usage{
rbmi_test_data
}
\description{
Example dataset for \code{tern.rbmi} package. This is an pool object from the rbmi analysis, see
\code{browseVignettes(package = "tern.rbmi")}
}
\keyword{datasets}
