% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_rbmi.R
\name{s_rbmi_lsmeans}
\alias{s_rbmi_lsmeans}
\title{Statistics function which is extracting estimates from a tidied LS means
data frame.}
\usage{
s_rbmi_lsmeans(df, .in_ref_col, show_relative = c("reduction", "increase"))
}
\arguments{
\item{df}{input dataframe}

\item{.in_ref_col}{boolean variable, if reference column is specified}

\item{show_relative}{should the "reduction" (\code{control - treatment}, default) or the "increase"
(\code{treatment - control}) be shown for the relative change from baseline?}
}
\description{
Statistics function which is extracting estimates from a tidied LS means
data frame.
}
