% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nesttemplate.R
\docType{package}
\name{tern.rbmi-package}
\alias{tern.rbmi}
\alias{tern.rbmi-package}
\title{tern.rbmi: Create TLGs using rbmi and tern}
\description{
\code{nesttemplate} provides a custom \href{https://pkgdown.r-lib.org}{\code{pkgdown}}
template for NEST packages. Please don’t use it for your own package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/tern.rbmi}
  \item \url{https://insightsengineering.github.io/tern.rbmi/}
  \item Report bugs at \url{https://github.com/insightsengineering/tern.rbmi/issues}
}

}
\author{
\strong{Maintainer}: NEST \email{basel.nestcicd@roche.com}

}
\keyword{internal}
