% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_rbmi.R
\name{tidy.pool}
\alias{tidy.pool}
\title{Helper method (for \code{\link[broom:reexports]{broom::tidy()}}) to prepare a data frame from an
\code{pool} \code{rbmi} object containing the LS means and contrasts and multiple visits}
\usage{
\method{tidy}{pool}(x)
}
\arguments{
\item{x}{(\code{pool}) is a list of pooled object from \code{rbmi} analysis results. This list includes
analysis results, confidence level, hypothesis testing type.}
}
\value{
A dataframe
}
\description{
Helper method (for \code{\link[broom:reexports]{broom::tidy()}}) to prepare a data frame from an
\code{pool} \code{rbmi} object containing the LS means and contrasts and multiple visits
}
