% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_rbmi.R
\name{h_tidy_pool}
\alias{h_tidy_pool}
\title{Helper function to produce data frame with results
of pool for a single visit}
\usage{
h_tidy_pool(x)
}
\arguments{
\item{x}{(\code{pool}) is a list of pooled object from \code{rbmi} analysis results. This list includes
analysis results, confidence level, hypothesis testing type.}
}
\value{
Data frame with results of pool for a single visit.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\examples{
data("rbmi_test_data")
pool_obj <- rbmi_test_data

h_tidy_pool(pool_obj$pars[1:3])

}
