% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tern.rbmi-package.R
\docType{package}
\name{tern.rbmi-package}
\alias{tern.rbmi}
\alias{tern.rbmi-package}
\title{tern.rbmi Package}
\description{
\code{tern.rbmi} is an analysis package to create tables, listings and graphs to analyze clinical trials data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/insightsengineering/tern.rbmi}
  \item \url{https://insightsengineering.github.io/tern.rbmi/}
  \item Report bugs at \url{https://github.com/insightsengineering/tern.rbmi/issues}
}

}
\author{
\strong{Maintainer}: Joe Zhu \email{joe.zhu@roche.com} (\href{https://orcid.org/0000-0001-7566-2787}{ORCID})

Authors:
\itemize{
  \item Jana Stoilova \email{jana.stoilova@roche.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
